--- src/favs/interfaces/favs.interfaces.ts
@@ -1,9 +1,0 @@
-import { Album } from 'src/album/entity/album.entity';
-import { Artist } from 'src/artist/entity/artist.entity';
-import { Track } from 'src/track/entity/track.entity';
-
-export interface FavoritesResponse {
-  artists: Artist[];
-  albums: Album[];
-  tracks: Track[];
-}
