+++ src/resources/artist/artist.module.ts
