+++ src/resources/auth/auth.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { JwtModule } from '@nestjs/jwt';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { User } from 'src/user/entity/user.entity';
+import { User } from 'src/resources/user/entity/user.entity';
 import { AuthController } from './auth.controller';
 import { AuthService } from './services/auth.service';
