+++ src/resources/favs/entity/favs.entity.ts
